+++ packages/common/qds-core/src/button/icon-button.api.ts
@@ -28,6 +28,9 @@ export function createQdsIconButtonApi(
   const variant = props.variant || "fill"
 
   return {
+    size,
+
+    // group: bindings
     getIconBindings(): QdsIconButtonIconBindings {
       return normalize.element({
         ...parts.icon,