+++ packages/frameworks/react-swagger/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @qualcomm-ui/react-swagger Changelog
 
+## 1.3.27
+
+Aug 6th, 2026
+
+### Miscellaneous Chores
+
+- **deps:** update dependencies [@qualcomm-ui/qds-core@1.34.0]
+
 ## 1.3.26
 
 Jul 28th, 2026